requirements.txt
# core
numpy
matplotlib
pillow
h5py
scikit-image
funcsigs
opencv-python
moviepy
# torch==1.9.0
# torchvision==0.10.0
# tensorboard==2.1.1
# tensorboardX==2.0
torch
torchvision
tensorboard
tensorboardX
gym==0.15.4
pandas

# RL
wandb
mpi4py
# mujoco_py==2.0.2.9
mujoco_py
